working with proper cleanup and new google voice

+++ src/ivrs_utils.py
@@ -4,7 +4,9 @@ def killtree(pid, including_parent=True):
     parent = psutil.Process(pid)
     for child in parent.children(recursive=True):
         print ("child", child)
-        child.kill()
+        childStatus = child.status()
+        if childStatus and childStatus != 'terminated':
+            child.kill()
 
     if including_parent:
         parent.kill()
